move data loading fxns to load.py
